requirements.txt
openai
deepgram-sdk
elevenlabs
python-dotenv
google-cloud-texttospeech
boto3